# api/openapi.yaml
openapi: 3.1.0
info:
  title: Samyama Graph Database API
  version: 1.8.0
  description: |
    HTTP API for the Samyama high-performance distributed graph database.
    Supports OpenCypher queries, graph status, and CRUD operations.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/query:
    post:
      operationId: executeQuery
      summary: Execute a Cypher query
      description: |
        Execute an OpenCypher query against the graph database.
        Supports both read (MATCH) and write (CREATE, SET, DELETE, MERGE) queries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            examples:
              matchAll:
                summary: Match all Person nodes
                value:
                  query: "MATCH (n:Person) RETURN n"
              matchWithFilter:
                summary: Match with WHERE filter
                value:
                  query: "MATCH (n:Person) WHERE n.age > 25 RETURN n.name, n.age"
              createNode:
                summary: Create a new node
                value:
                  query: 'CREATE (n:Person {name: "Alice", age: 30})'
              traversal:
                summary: Traverse relationships
                value:
                  query: "MATCH (a:Person)-[:KNOWS]->(b:Person) RETURN a.name, b.name"
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Query error (parse error, execution error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/begin:
    post:
      operationId: beginTransaction
      summary: Begin a transaction
      description: |
        Opens a transaction and returns its id. Send statements to `/api/query`
        with `"tx": <id>`, then commit or roll back.

        The transaction holds the graph's write lock until it ends: no other
        request can read or write while it is open. It is rolled back
        automatically after `timeout_seconds` (`SAMYAMA_TX_TIMEOUT_SECS`,
        default 30). Isolation is serializable in effect, since nothing runs
        concurrently with an open transaction.
      responses:
        '200':
          description: Transaction opened
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  version:
                    type: integer
                  timeout_seconds:
                    type: integer
        '409':
          description: A transaction could not be opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/{id}/commit:
    post:
      operationId: commitTransaction
      summary: Commit a transaction
      description: Keeps every write the transaction made, persists them, and releases the lock.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Committed
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  committed:
                    type: boolean
                  version:
                    type: integer
        '404':
          description: No such open transaction (already ended, or timed out and rolled back)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The transaction timed out and was rolled back
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tx/{id}/rollback:
    post:
      operationId: rollbackTransaction
      summary: Roll back a transaction
      description: |
        Undoes everything the transaction did -- property and label writes,
        created nodes and relationships, and deletions, which are restored with
        their properties and relationships -- and releases the lock.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rolled back
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx:
                    type: string
                  rolled_back:
                    type: boolean
        '404':
          description: No such open transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/import/csv:
    post:
      operationId: importCsv
      summary: Import nodes from CSV
      description: |
        Upload a CSV file to create nodes. The first row is treated as column headers
        (property names). Each subsequent row creates one node with the specified label.
        Property types are auto-detected (integer → float → boolean → string).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - label
              properties:
                file:
                  type: string
                  description: CSV file content
                label:
                  type: string
                  description: Node label to assign to all created nodes
                id_column:
                  type: string
                  description: Column to use as node ID mapping (optional)
                delimiter:
                  type: string
                  description: "CSV delimiter character (default: ',')"
                graph:
                  type: string
                  description: "Target graph/tenant name (default: 'default')"
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsvImportResponse'
        '400':
          description: Import error (missing fields, parse error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/import/parquet:
    post:
      operationId: importParquet
      summary: Import nodes from a Parquet file
      description: |
        Upload a Parquet file to create one node per row, with the file's columns as
        properties. Nodes only: an edge needs identity resolution — a `source` column
        names *something*, and which property it names is a decision the file cannot
        make — so that belongs in `LOAD PARQUET`, where the mapping is written in the
        query.
      parameters:
        - name: label
          in: query
          required: true
          schema:
            type: string
          description: |
            The label every created node gets. Required rather than inferred: a Parquet
            file carries a schema, not a label, and taking one from the filename would
            make the import depend on what the file was called.
        - name: graph
          in: query
          required: false
          schema:
            type: string
            default: default
          description: Target graph/tenant name
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Parquet file content
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    examples:
                      - ok
                  stats:
                    type: object
                    description: Rows read and nodes created
        '400':
          description: Missing file field, unreadable Parquet, or unknown graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/query/export:
    post:
      operationId: exportQuery
      summary: A read query's result as Arrow IPC or Parquet
      description: |
        Runs a read query and returns the result as a binary columnar stream rather
        than JSON. `arrow` is the default because it needs no intermediate file at
        either end.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: The Cypher read query
                graph:
                  type: string
                  default: default
                  description: Target graph/tenant name
                format:
                  type: string
                  enum: [arrow, parquet]
                  default: arrow
                  description: Output encoding
      responses:
        '200':
          description: The result set
          content:
            application/vnd.apache.arrow.stream:
              schema:
                type: string
                format: binary
            application/vnd.apache.parquet:
              schema:
                type: string
                format: binary
        '400':
          description: Unknown format, unknown graph, or a query that does not run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/import/json:
    post:
      operationId: importJson
      summary: Import nodes from JSON
      description: |
        Create nodes from an array of JSON objects. Each object becomes one node
        with the specified label. Object keys become property names; values must be
        string, number, or boolean (other types are silently skipped).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonImportRequest'
            example:
              label: Person
              nodes:
                - name: Alice
                  age: 30
                - name: Bob
                  age: 25
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonImportResponse'
        '400':
          description: Import error (missing label, invalid data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/sample:
    post:
      operationId: sampleSubgraph
      summary: Sample a representative subgraph
      description: >
        Returns a label-stratified, size-capped subgraph intended for
        visualization. One call replaces the `1 + <edge types>` sequential
        queries a client would otherwise issue to draw an overview, and the
        default of 200 nodes is the point past which a node-link diagram stops
        being legible.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                max_nodes:
                  type: integer
                  default: 200
                  description: Upper bound on returned nodes. Capped server-side at 1000.
                labels:
                  type: array
                  items:
                    type: string
                  description: Restrict sampling to these labels. Empty means all labels.
                graph:
                  type: string
                  description: Tenant/graph name.
      responses:
        '200':
          description: A sampled subgraph
          content:
            application/json:
              schema:
                type: object
  /api/vector/indexes:
    get:
      operationId: listVectorIndexes
      summary: List vector indexes
      responses:
        '200':
          description: The configured vector indexes
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: createVectorIndex
      summary: Create a vector index
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - label
                - property_key
                - dimensions
              properties:
                label:
                  type: string
                property_key:
                  type: string
                dimensions:
                  type: integer
                metric:
                  type: string
                  description: Distance metric, e.g. cosine or euclidean.
      responses:
        '200':
          description: The index was created
  /api/vector-search:
    post:
      operationId: vectorSearch
      summary: k-nearest-neighbour search over a vector index
      description: >
        Supply either `query_vector` or `query_text`; supplying text requires an
        embedding provider to be configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query_text:
                  type: string
                query_vector:
                  type: array
                  items:
                    type: number
                    format: float
                label:
                  type: string
                property_key:
                  type: string
                k:
                  type: integer
                  description: Number of neighbours to return.
      responses:
        '200':
          description: The nearest neighbours, closest first
          content:
            application/json:
              schema:
                type: object
...
  /api/snapshot/import:
    post:
      operationId: importSnapshot
      summary: Import a .sgsnap snapshot into a tenant
      description: >
        The body is **multipart/form-data**, not a raw octet stream — posting the
        bytes directly is the mistake this description exists to prevent. Import
        merges into the target tenant rather than replacing it.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: The snapshot was imported
...
components:
  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: An OpenCypher query string
          examples:
            - "MATCH (n:Person) RETURN n"
        graph:
          type: string
          description: "Target graph/tenant name (default: 'default')"
          default: default
        cache:
          type: boolean
          description: >-
            Serve this query from the result cache when the graph has not
            changed since the answer was computed. Omitted means the server
            default, which is off unless SAMYAMA_RESULT_CACHE=1. Send false to
            bypass a warm cache for a latency measurement.
        tx:
          type: string
          description: >-
            Run the statement inside this open transaction, from
            `POST /api/tx/begin`. Its writes are visible only inside the
            transaction until `POST /api/tx/{id}/commit`, and
            `POST /api/tx/{id}/rollback` undoes them.
...
    JsonImportRequest:
      type: object
      required:
        - label
        - nodes
      properties:
        label:
          type: string
          description: Node label to assign to all created nodes
        nodes:
          type: array
          description: Array of JSON objects, each becoming a node
          items:
            type: object
            additionalProperties: true
        graph:
          type: string
          description: "Target graph/tenant name (default: 'default')"
          default: default
    CsvImportResponse:
      type: object
      properties:
        status:
          type: string
          examples:
            - ok
        nodes_created:
          type: integer
        label:
          type: string
        columns:
          type: array
          items:
            type: string
    JsonImportResponse:
      type: object
      properties:
        status:
          type: string
          examples:
            - ok
        nodes_created:
          type: integer
        label:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...